@@ -1,54 +1,40 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import { FlatList } from 'react-native';
 import { getDecks } from '../../utils/api';
 import Deck from '../../components/Deck/Deck';
 
-class DeckList extends React.PureComponent {
-  state = {
-    decks: null
-  }
+const DeckList = ({ navigation }) => {
+  const [decks, setDecks] = useState(null);
 
   // Refresh the list of decks
-  refreshDecks = async () => {
-    const decks = await getDecks();
-    this.setState({ decks });
-  }
+  const refreshDecks = async () => {
+    const decksData = await getDecks();
+    setDecks(decksData);
+  };
 
   // Initialize the deck list
-  componentDidMount() {
-    this.refreshDecks();
-  }
-
-  // Reload the deck list upon adding a deck
-  componentDidUpdate(prevProps, prevState) {
-    if (prevState !== this.state) {
-      this.refreshDecks();
-    }
-  }
+  useEffect(() => {
+    refreshDecks();
+  }, []);
 
   // Display details for an individual deck when that deck is pressed
-  handlePress = (deck) => {
-    const { navigation } = this.props;
-    navigation.navigate('DeckDetail', { deck });
+  const handlePress = (deck) => {
+    navigation.navigate('DeckDetail', { deck, deckId: deck.title });
+  };
+
+  if (decks) {
+    const deckArray = Object.keys(decks).map(deckName => decks[deckName]); // Format decks for <FlatList /> component
+
+    return (
+      <FlatList
+        data={deckArray}
+        renderItem={({ item }) => <Deck deck={item} onPress={() => handlePress(item)} />}
+        keyExtractor={item => item.title}
+      />
+    );
   }
 
-  render() {
-    const { decks } = this.state;
-
-    if (decks) {
-      const deckArray = Object.keys(decks).map(deckName => decks[deckName]); // Format decks for <FlatList /> component
-
-      return (
-        <FlatList
-          data={deckArray}
-          renderItem={({ item }) => <Deck deck={item} onPress={() => this.handlePress(item)} />}
-          keyExtractor={item => item.title}
-        />
-      );
-    }
-
-    return null;
-  }
-}
+  return null;
+};
 
 export default DeckList;
