@@ -1,53 +1,48 @@
-import React from 'react';
+import React, { useLayoutEffect } from 'react';
 import { View, Text } from 'react-native';
 import styles from './styles';
-import { getDeck } from '../../utils/api';
 import { determineCardPlurality } from '../../utils/helpers';
 import TextButton from '../../components/TextButton/TextButton';
 
-class Score extends React.PureComponent {
-  // Set screen header
-  static navigationOptions = ({ navigation }) => {
-    const { deck } = navigation.state.params;
+const Score = ({ route, navigation }) => {
+  const { deck, numCorrect } = route.params;
 
-    return ({
+  // Set screen header
+  useLayoutEffect(() => {
+    navigation.setOptions({
       title: `${deck.title} Quiz Score`
     });
-  }
+  }, [navigation, deck]);
 
-  render() {
-    const { deck, numCorrect } = this.props.navigation.state.params;
-
-    if (deck) {
-      return (
-        <View style={styles.container}>
-          <View className="text" style={styles.subcontainer}>
-            <Text style={styles.heading}>{`${100.0 * numCorrect / deck.questions.length}%`}</Text>
-            <Text style={styles.caption}>{`You got ${numCorrect} / ${deck.questions.length} ${determineCardPlurality(deck)} correct!`}</Text>
-          </View>
-          <View className="buttons" style={styles.subcontainer}>
-            <TextButton 
-              text="Restart Quiz" 
-              onPress={() => this.props.navigation.navigate('Quiz', {
-                deck,
-                questionIndex: 0,
-                numCorrect: 0
-              })} 
-            />
-            <TextButton 
-              text="Back to Deck" 
-              onPress={() => this.props.navigation.navigate('DeckDetail', {
-                deck
-              })} 
-            />
-          </View>
+  if (deck) {
+    return (
+      <View style={styles.container}>
+        <View style={styles.subcontainer}>
+          <Text style={styles.heading}>{`${(100.0 * numCorrect / deck.questions.length).toFixed(0)}%`}</Text>
+          <Text style={styles.caption}>{`You got ${numCorrect} / ${deck.questions.length} ${determineCardPlurality(deck)} correct!`}</Text>
         </View>
-      );
-    }
-
-    return null;
-    
+        <View style={styles.subcontainer}>
+          <TextButton
+            text="Restart Quiz"
+            onPress={() => navigation.navigate('Quiz', {
+              deck,
+              questionIndex: 0,
+              numCorrect: 0
+            })}
+          />
+          <TextButton
+            text="Back to Deck"
+            onPress={() => navigation.navigate('DeckDetail', {
+              deck,
+              deckId: deck.title
+            })}
+          />
+        </View>
+      </View>
+    );
   }
-}
+
+  return null;
+};
 
 export default Score;
